fix: notification class and extensions
