+++ Water Reminder/Water Reminder/Extensions/String+ToDouble.swift
@@ -1,0 +1,14 @@
+//
+//  String+ToDouble.swift
+//  Water Reminder
+//
+//  Created by Дарья Воробей on 11/22/21.
+//
+
+import Foundation
+
+extension String {
+    func toDouble() -> Double? {
+        return NumberFormatter().number(from: self)?.doubleValue
+    }
+}
